+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # kotlin-code-sort Changelog
 
 ## [Unreleased]
+### Added
+- Integration with IntelliJ's code cleanup feature, allowing Kotlin code to be sorted during code cleanup
 
 ## [0.1.0]
 - Add settings to customize the sort order, sort alphabetically or exclude certain elements from sorting